Crate invoker_api

Modules

judge_log
Judge log stored in FS

status_codes

valuer_proto
Defines types used to interact between invoker and valuer

Structs

CliInvokeTask
Pass this to invoker running in CLI mode

DbInvokeTask
Put this in database

InvokeOutcomeHeader

InvokeTask

LiveStatusUpdate
Represents Live Status Update. Some fields can be None always, or only in some updates.

Status

Enums

StatusKind